# satellite/satellitedb: ensure that we process orders in order