--- a/pkg/storj/node.go
+++ b/pkg/storj/node.go
@@ -110,7 +110,7 @@ func (id NodeID) IsZero() bool {
 // Bytes returns raw bytes of the id
 func (id NodeID) Bytes() []byte { return id[:] }
 
-// Less returns whether id is smaller than other in lexicographic order
+// Less returns whether id is smaller than other in lexicographic order.
 func (id NodeID) Less(other NodeID) bool {
 	for k, v := range id {
 		if v < other[k] {