--- a/pkg/storj/serialnumber.go
+++ b/pkg/storj/serialnumber.go
@@ -44,7 +44,7 @@ func (id SerialNumber) IsZero() bool {
 	return id == SerialNumber{}
 }
 
-// Less returns whether id is smaller than other in lexicographic order
+// Less returns whether id is smaller than other in lexicographic order.
 func (id SerialNumber) Less(other SerialNumber) bool {
 	for k, v := range id {
 		if v < other[k] {