--- a/satellite/satellitedb/orders.go
+++ b/satellite/satellitedb/orders.go
@@ -239,7 +239,7 @@ func (db *ordersDB) ProcessOrders(ctx context.Context, requests []*orders.Proces
 	storageNodeID := requests[0].OrderLimit.StorageNodeId
 	for _, req := range requests[1:] {
 		if req.OrderLimit.StorageNodeId != storageNodeID {
-			return nil, Error.New("requests from different different storage nodes %v and %v", storageNodeID, req.OrderLimit.StorageNodeId)
+			return nil, Error.New("requests from different storage nodes %v and %v", storageNodeID, req.OrderLimit.StorageNodeId)
 		}
 	}