What: When transactions are executed in random order concurrently then the postgres database can lockup. Making the transaction fail with a message "database locked".

As a bonus there is a performance improvement:
```
benchmark                                      old ns/op      new ns/op     delta
BenchmarkOrders/Sqlite/Process:50-32           7266013        4679146       -35.60%
BenchmarkOrders/Postgres/Process:50-32         38178964       39148530      +2.54%
BenchmarkOrders/Sqlite#01/Process:100-32       17596919       9265638       -47.35%
BenchmarkOrders/Postgres#01/Process:100-32     79357764       75785671      -4.50%
BenchmarkOrders/Sqlite#02/Process:250-32       70267380       22843365      -67.49%
BenchmarkOrders/Postgres#02/Process:250-32     187335133      177281750     -5.37%
BenchmarkOrders/Sqlite#03/Process:500-32       218870500      45772104      -79.09%
BenchmarkOrders/Postgres#03/Process:500-32     449023067      394666333     -12.11%
BenchmarkOrders/Sqlite#04/Process:999-32       717762000      90001023      -87.46%
BenchmarkOrders/Postgres#04/Process:999-32     1110214300     799247200     -28.01%
```

Why:

Please describe the tests:
 - Test 1:
 - Test 2:
 
Please describe the performance impact:

## Code Review Checklist (to be filled out by reviewer)
 - [x] Does the PR describe what changes are being made?
 - [x] Does the PR describe why the changes are being made?
 - [ ] Does the code follow [our style guide](https://github.com/storj/docs/blob/master/code/Style.md)?
 - [ ] Does the code follow [our testing guide](https://github.com/storj/docs/blob/master/code/Testing.md)?
 - [x] Is the PR appropriately sized? (If it could be broken into smaller PRs it should be)
 - [ ] Does the new code have enough tests? (*every* PR should have tests or justification otherwise. Bug-fix PRs especially)
 - [ ] Does the new code have enough documentation that answers "how do I use it?" and "what does it do?"? (both source documentation and [higher level](https://github.com/storj/docs), diagrams?)
 - [x] Does any documentation need updating?
 - [x] Do the database access patterns make sense?