@stefanbenten 2019-09-04
We could use a test to ensure this type of error is never happening again :)
Maybe @ethanadams has an easy test to add?