@littleskunk 2019-09-04
Negative. Deadlock still in place.
```
satellite/peer.go:722	Node 18zEcpEtGjwtXT9tBbJnKPfyGC3P3ZbwFZkzEyhiYajtmzRBYS started
[...]
storagenode/1    12gPTSsJdz 21:55:22.124 | ERROR	orders.18zEcpEtGjwtXT9tBbJnKPfyGC3P3ZbwFZkzEyhiYajtmzRBYS	orders/service.go:194	failed to settle orders	{"error": "order: failed to receive settlement response: rpc error: code = Unknown desc = pq: deadlock detected", "errorVerbose": "order: failed to receive settlement response: rpc error: code = Unknown desc = pq: deadlock detected\n\tstorj.io/storj/storagenode/orders.(*Service).settle:312\n\tstorj.io/storj/storagenode/orders.(*Service).Settle:192\n\tstorj.io/storj/storagenode/orders.(*Service).sendOrders.func2:171\n\tgolang.org/x/sync/errgroup.(*Group).Go.func1:57"}
                                         | storj.io/storj/storagenode/orders.(*Service).Settle
                                         | 	/home/newrelease/storj/storagenode/orders/service.go:194
                                         | storj.io/storj/storagenode/orders.(*Service).sendOrders.func2
                                         | 	/home/newrelease/storj/storagenode/orders/service.go:171
                                         | golang.org/x/sync/errgroup.(*Group).Go.func1
                                         | 	/home/newrelease/go/pkg/mod/golang.org/x/sync@v0.0.0-20190423024810-112230192c58/errgroup/errgroup.go:57
```